BASE-276: - renamed chart2_wdg to chart_wdg and changed chart_wdg to chart_wdg_old to prevent import issues - also renamed all instances of the above
