+++ src/tactic/ui/chart/__init__.py
@@ -11,9 +11,8 @@
 #
 
 #from chart_data import *
-#from chart_builder_wdg import *
-#from chart_wdg import *
-from chart2_wdg import *
+from chart_builder_wdg import *
+from chart_wdg import *
 
 #from bar_chart_wdg import *
 #from date_chart_wdg import *
